third_party/WebKit/LayoutTests/fast/multicol/balance-float-in-inline.html

<!DOCTYPE html>
<p>There should be a blue square below.</p>
<div style="position:relative; columns:4; column-gap:0; width:600px;">
    <br>
    <span>
        <div id="float" style="float:right; break-inside:avoid; width:150px; height:150px; background:blue;"></div>
    </span>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
    var elm = document.getElementById("float");
    assert_equals(elm.offsetTop, 0);
    assert_equals(elm.offsetLeft, 150);
    assert_equals(elm.offsetWidth, 150);
    assert_equals(elm.offsetHeight, 150);
}, "Unbreakable float inside inline.");
</script>
